fix: header files and code maintainability
