+++ include/email.h
@@ -1,0 +1,7 @@
+#pragma once //prevent file from being included more than once
+
+class MAILRESULTS{
+private:
+public:
+void send();
+};
