+++ include/settings.h
@@ -1,0 +1,12 @@
+#pragma once //prevent file from being included more than once
+
+// The smtp host name e.g. smtp.gmail.com for GMail or smtp.office365.com for Outlook or smtp.mail.yahoo.com
+#define SMTP_HOST "smtp.gmail.com"
+#define SMTP_PORT 465
+
+// The sign in credentials
+#define AUTHOR_EMAIL "open.vibration.analysis@gmail.com"
+#define AUTHOR_PASSWORD "uiea whvy jzve qplh"
+
+// Recipient's email
+#define RECIPIENT_EMAIL "mvanessen15@gmail.com"
